requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
paramiko==3.4.0
python-socketio==5.10.0
python-engineio==4.8.0
Werkzeug==3.0.1
requests==2.31.0
psutil==5.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
